refactor: move logErrorMessage into loggerUtils
