+++ backend/scripts/populateMessagesScript.ts
@@ -3,7 +3,7 @@ import {
   getSeedCommunitiesWithOwnerAndInbox,
   createCommunitySeedMessage,
 } from "../database/prisma/scriptQueries";
-import { logErrorMessage } from "../utils/error";
+import { logErrorMessage } from "../utils/loggerUtils";
 import {
   logPopulateMessage,
   logPopulateSuccessMessage,
