+++ backend/scripts/populateNotificationsScript.ts
@@ -4,7 +4,7 @@ import {
   createCommunityNotification,
   getSeedUsersWithLimit,
 } from "../database/prisma/scriptQueries";
-import { logErrorMessage } from "../utils/error";
+import { logErrorMessage } from "../utils/loggerUtils";
 import { CommunityWithOwnerAndInbox } from "../@types/scriptTypes";
 import {
   logPopulateMessage,
