+++ backend/scripts/unpopulateDB.ts
@@ -9,7 +9,7 @@ import {
   getSeedLoneUsers,
   getSeedUsers,
 } from "../database/prisma/scriptQueries";
-import { logErrorMessage } from "../utils/error";
+import { logErrorMessage } from "../utils/loggerUtils";
 import {
   logUnpopulateMessage,
   logUnpopulateSuccessMessage,
