+++ backend/scripts/unpopulateMessagesDB.ts
@@ -3,7 +3,7 @@ import {
   deleteCommunityMessages,
   getSeedCommunities,
 } from "../database/prisma/scriptQueries";
-import { logErrorMessage } from "../utils/error";
+import { logErrorMessage } from "../utils/loggerUtils";
 import {
   logUnpopulateMessage,
   logUnpopulateSuccessMessage,
