+++ backend/utils/error.ts
@@ -15,13 +15,4 @@ const createValidationErrObj = (err: Result, message: string) => {
   };
 };
 
-const logErrorMessage = (
-  // eslint-disable-next-line
-  error: any,
-) => {
-  if (error instanceof Error) {
-    console.log(error.message);
-  }
-};
-
-export { createValidationErrObj, logErrorMessage };
+export { createValidationErrObj };
